Put properties of Newton method into separate header for reuse.
Now they can be reused in opm-simulators without any hassle.
